videomae/requirements.txt
timm==0.4.12
deepspeed==0.7.3
tensorboardX
decord
einops
opencv-python
scipy
pandas
pytorch-lightning>=2.0.0
pyyaml
schedulefree
torchvision
numpy==1.26.4
av
litdata
optuna
plotly
kaleido
zstd
torchmetrics